backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic-settings==2.7.*
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
alembic==1.14.*
httpx==0.28.*
python-dotenv==1.0.*
openai>=1.0
anthropic>=0.40
google-genai>=1.0
